added form validation, created monster in process
